generify element list widgets and prevent wrapping soft-lock
